Small optimization: _PyDict_GetItem_KnownHashKnownDict
The _PyDict_GetItem_KnownHash function spends a decent amount of
its time checking if the argument is a dict, but most of the time
the caller knows that its a dict, so add a new function that
skips this check.

An interesting question is if we should do these sorts of internal
error checks at all.