Regenerate .pyc files after package installation The files contain the full path of installation directory used during the build e.g. /home/marius/dev/pyston_v2/pyston/build/cpython_unopt_install/usr/lib/python3.8-pyston2.3/re.py and as pointed out in #70 also the wrong timestamp. This means the common use-case where the package is installed by root and a different user is executing pyston can't use the .pyc files and instead has to parse the files on every import.
The ubuntu cpython package contains a similar script but which does not use `compileall.py` but instead
manually does most of the work and is not regenerating them for all optimization levels.
```bash
files=$(dpkg -L lib@PVER@-testsuite | sed -n '/^\/usr\/lib\/@PVER@\/.*\.py$/p' | egrep -v '/lib2to3/tests/data|/test/bad')
    if [ -n "$files" ]; then
        /usr/bin/@PVER@ -E -S /usr/lib/@PVER@/py_compile.py $files
        if grep -sq '^byte-compile[^#]*optimize' /etc/python/debian_config; then
            /usr/bin/@PVER@ -E -S -O /usr/lib/@PVER@/py_compile.py $files
        fi
    else
        echo >&2 "@PVER@: can't get files for byte-compilation"
    fi
```
I decided against copying the code and instead decided to stick to what cpython does in `Makefile.pre.in`
using `compileall.py`.

This does not change things for the portable package but I think here it's not a problem
because the directory will likely be writable by the user so pyston will just update
them the first time they are imported.

In addition added a script to remove the .pyc files when removing the package because dpkg will only remove the
files included in the package and will complain that it can't remove the directories if they contain newly created .pyc files.

Fixes #70