portable release: modify rpath of copied libraries We also have to modify the rpath of libraries we have copied to look inside the current directory. Before we only modified the rpaths of the python extensions inside port_rel/usr/lib/python3.8-pyston2.3/lib-dynload/ but not the files inside port_rel/usr/lib/.
E.g. port_rel/usr/lib/python3.8-pyston2.3/lib-dynload/readline.pyston-23-x86_64-linux-gnu.so
has dependency on libreadline.so.8 which we provide inside port_rel/usr/lib/
but it has a dependency on libtinfo.so.6 which may not be installed but we provide too.
But before it did not get picked up because the rpath of libreadline.so.8 did not take into account the current directory.

fixes #82 - I confirmed that now the arrow keys/'import readline' works on ubuntu 21.04 and fedora