Fix: need to change the stack limit here
I had assumed that PyThreadState objects were pinned to a specific system thread,
but that is not the case.

So hook the place that a new PyThreadState is set up for the system thread,
and update the stack limit there.

This is good evidence that this should be stored in (p)thread-local storage, not
a PyThreadState