Makefile fix: update this timestamp
We have a rule where executing the rule doesn't always update the timestamp of the target file. So explicitly update it to make sure that we don't keep having to run the rule.

The rule is
cpython_unopt_install/usr/lib/libpython3.8-pyston2.3.so.1.0 => unopt_env/bin/python

Which is added as a dependency in the bolt-so section