Link pools into generational lists
based on their youngest member. So if any object in a pool is
in generation 0, the pool is a member of generations[0]

Rename nextpool/prevpool to nextusedpool/nextusedpool now
that we have multiple linked lists a pool can be part of.

This commit starts by just linking pools into generation 0
when they have anything allocated from them. Scanning and generation
management to come.