conda: set python prefix to the conda prefix
Previously we set the prefix to /usr, then moved the files
to the appropriate prefix.

By configuring python with the proper conda prefix, conda can
do its prefix-rewriting when it installs.

This is usually not necessary since prefix-detection is usually done
via argv[0], but it caused some corner cases to fail as evidenced
by a failing test case.