aot: merge function and method The only difference was that function traces assumed that the oparg argument stays the same while for CALL_METHOD they can change.
This will add one if check which will always be false in exchange
for a reduction in then number of traces.
I think this is a good trade-off. When I added the differentiation between
call_function and call_method I did not realize that e.g. calling a function inside another module will use the CALL_METHOD
that's why I thought it would be useful to have different traces.