jit: introduce new jit for call argument setup this JIT generates code for CALL_FUNCTION, METHOD_CALL and CALL_FUNCTION_KW opcodes. The idea is to reduce the time spend inside _PyEval_EvalCodeWithName() by generating machine code for it. By guarding on the function version we know that the default args, code object etc can't change. This allows us to directly emit default args and kwargs (before it had to loop on every call to retrieve position etc..). With this change the following calls are all the same speed: def f(a, b, c=1, d=2): pass f(1, 2, 3, 4) f(1, 2) f(a=1, b=2)
In addition this removes one inc/decref per arg because the function on the python value stack are already owned
so we can copy it into the frameobject directly.

In the future we can add support for more code object types.

JIT_CALLSITES=1 enables this JIT (default: on)
JIT_CALLSITES_TRIVIAL=1 JIT calls with no default args and no kwargs (default: off - normal path has nearly same speed)

The JIT is called from the AOT Profile functions. I tried before to profile via opcache mechanism in the interpreter and
emit the callsite jit as an IC directly in the calling function but this had a serious issue:
The function object in the cache is only borrowed so when JITing it could have been already destroyed and we crashed.