Pin our dependencies in the same way the conda-forge python feedstock does
More importantly than setting the version, pinning them in this way
automatically adds them as run dependencies.

This also sets openssl to v1.1.1 which is important because it looks
like the conda-forge ecosystem is built on that version.
It also sets the compiler version which should avoid some other problems.