Constant-fold more comparisons with constants
For example if we know that an object is of type PyLong_Type,
and we check if it's equal to Py_False, we know it must be
not equal.

In general it eliminates any equality checks where we know conflicting
facts about the two operands.

Right now it checks Facts against JitConsts data, but maybe in the future
it would be great to integrate the JitConsts data into the Fact system.