Implement CALL_METHOD "hinting"
We will often have good runtime information available about
the callable we will call with CALL_METHOD, since this value
will often by cached for the corresponding LOAD_METHOD call.

This information is more specific than what our aot traces guards on,
so if we can remember it we can create an inline cache that is
quite a bit better than the aot trace. (For example we can statically
resolve getting the vectorcall function, type checking, and arg number checking.)

I was originally hoping to fuse the LOAD_METHOD + CALL_METHOD together,
which makes sense to me because then we can directly flow the info
within the same guard section.
But since there can be many bytecodes in between the LOAD_METHOD and CALL_METHOD,
just store some hint information along with the guard that's required to
validate the hint.

todo:
- LOAD_GLOBAL + CALL_FUNCTION can probably also benefit from this.
- It feels like a more general system could probably pick up a lot
  of similar cross-bytecode opportunities like this.
- There is still a decent amount of overhead left that we could get rid
  of. For example shuffling things to-and-from the value stack,
  incref+decreffing pairs that are unnecessary, etc.
  - I think this might already be possible with smarter use of the deferred_vs functions