Argument parsing: fast-path this check
PyFloat_Check is somewhat expensive because it doesn't use the fast
bitmask method. And presumably this error case (passing a float to
a function expecting a long) is rare, and the much-more-common case
is we pass a long to these functions.

So first check if the type is long, and if it is then check the expensive
PyFloat_Check check.