Properly free the hints
I added code to free them all at the end, but forgot that I
pop them as they are used. So free them after they are popped.

Also add some defensive coding in case we are run on improper bytecode
where the LOAD_METHOD and CALL_METHOD are not matched.