Better test process isolation for test_django.py
In certain circumstances the django test runner will write
files into whatever the current directory is.

In particular it seems to sometimes write out a "cache/" directory.
And then during test discovery, if there is a "cache/" directory in
the working directory, that directory will shadow the "cache/" directory
of tests and they won't be discovered.