Generate the unspecialized cases first
This is a bit of a workaround

The problem is that in Orc V2, or at least the way I'm using it in this changeset,
resolving a symbol adds it to the symbol table. Which means that if we later try
to JIT a function with that name, it will fail with a "duplicate symbol name" error.

This happens if we trace ex call_functionisinstanceLong3 first, because it will
reference the partially-specialized call_functionisinstance3. Then when we go to
create the lesser specialization it will fail.

My approach here is to just generate the partial specializations first.