Fix: we were over-relying on llvm types
In a couple places we make decisions based on the llvm type, and
in this case whether the type is a pointer or an integer.

But in this new version of llvm it's loading a function pointer as
an int (and then immediately bitcasting it to a pointer). The previous
logic just thought that we happened to mean the constant value 0x4643d0,
and embedded that directly in the IR, meaning the binary hardcoded the
address that this function was at during tracing.

Now treat all integers as potential relocatable symbols