Give an aliasing hint to the trace optimizer
Some of the call_function traces (I'm looking specifically at call_functionlenList2) have redundant loads from the args array. This is especially detrimental because the repeated loads cause us to lose the type information that we had established via the guards, so the len() traces that are specialized on their argument type were essentially unspecialized.

Adding the "restrict" keyword here seems to be sufficient to teach llvm that the array is only mutated if we mutate it.

I had an earlier version of this change where I taught the optimizer that this array is constant, but this caused segfaults -- turns out in certain cases this array is allowed to be modified. So I went with "restrict" since it seems like the actual semantics we want.
(The reason I went with the earlier approach was LLVM 10 was unable to reason about the inline assembly we use for recursion checking and "restrict" was insufficient, but now with LLVM 13 this is enough.)