Fix: make sure to try specializations from most-specialized to least
We used to do this, but when I reordered the case generation it also
changed the order in which the specializations were tried at runtime.

Add a very simple heuristic of which cases are more specialized than
others, and test them from most-specialized to least.

The heuristic doesn't really distinguish between unrelated cases, it
just makes sure that len(list) is tried before len(object) is tried
before CFunction()