Immortalize the initial C modules
After playing around with it a bit, this is about as much as
I think can be immortalized without causing issues:
- initial C modules, their transitively-referenced modules, and all their attributes of the following types: CFunction, Unicode, Long, and Function

immortalizing the python modules as well immortalized too much.

immortalizing everything referenced by the initial C modules seemed to work,
but seems a bit scary to me, so I limited it just to a few types.