use compiler builtin math functions I tracked down why the nbody.py benchmark is slower with the conda build pyston than 'regular' one on ARM64. It comes down that it does not inline many math functions like isinf, isnan,...
Here is an example for isnan (but isinf etc are same):
int foo(double d) {
    return isnan(d);
}

gcc/clang on ubuntu 20.04:
0000000000000000 <foo>:
   0:	1e602000 	fcmp	d0, d0
   4:	1a9f77e0 	cset	w0, vs
   8:	d65f03c0 	ret

conda-forge gcc-11:
0000000000000000 <foo>:
   0:	14000000 	b	0 <__isnan>

Problem is the use of a ancient glibc 2.17 from 2012? which does not use the __builtin_isnan functions but always implements it themself.
Seems like in 2015 they started to use the compiler builtins.
https://github.com/lattera/glibc/commit/8df4e219e43a4a257d0759b54fef8c488e2f282e

We tithly control which compiler gets used when building Pyston so I think we can safely
use the builtin functions.

This makes the performance much more equal when running nbody.py.