docker: create multi arch images using buildx This seems to be the recommended way to create this images.
I tested a dummy release on my private repo (with test packages so please don't use this for anything) and
could push it successfully to undingen/pyston.
Afterwards I docker run it fine on my x86 PC and MacOS M1 ARM.

I decided to use the debian arch names (amd64,arm64) as the suffix for the debian packages.
And the conda default ones (x86_64,aarch64) for the installers like e.g. similar to the miniconda installer.