ICs: add polymorphic LOAD_ATTR/LOAD_METHOD caches Add a new cache type which stores a dynamically allocated array to further opcache entries. This are mostly useful for the JIT which can emit several entries with low overhead.
I had to come up with some heuristics to not generate too many polymorphic sites
and what seemed to work was to check if the type is the same (we us a hash because
of storage requirements in the cache entry). This seems to catch a lot of cases
where the shape is quickly changing. In additon the JIT will only emit entries
which at least once got executed which also catches some bad entries.

This improves our benchmarks but also pyperformance (especialy 'richards': 1.39x faster)

Cache entries will never be freed so will leak but I this is similar
to how JIT memory is already leaking.