jit: STORE_ATTR inline cache for split dicts Unfortunately it's not really possible to inline all of it because of the complexity of setting a dict entry so we still branch out to a helper function at the end.
But it's a bit faster then not having the cache inline.
make measure perf stays the same microbenchmarks are slightly improved (e.g richards).
Cache hit rates are really really good for most benchmarks I looked at.

I hope that in the future with the cpython 3.11 object layout for split dicts this caches
should be faster and smaller because the pointer to the values array is directly stored
in front of the object.