add some #line directives to the dynasm preprocessor
I'm seeing gdb display the wrong line numbers for things, and I think
what's going on is that #line directives inside a non-taken #if block are
ignored.

So when dynasm inserts some lines inside an #if block as well as a #line
directive, the #line directive is ignored but the extra lines are still counted,
resulting in too-large of a line number.

So this change adds extra #line directives after preprocessor directives.

My understanding is this issue was present pre-pyston-lite, we just didn't
see it until pyston-lite added a number of #if blocks.

It also adds line numbers after any dynasm directives so that all #line directives
point at the original file -- for consistency it's nice if they point to the same one,
and I think it's nicer if they point to the original so that you can do things like
"break aot_ceval_jit.c:2408" instead of "break aot_ceval_jit.prep.c:2408"