Better .pth installation
The stackoverflow answer I copied was wrong, so I copied how setuptools does it.

Keep the .pth file so that we can bootstrap the pgo process when we run the
pyston-lite setup.py: we're not able to run the autoloader setup.py yet
at that point, so we install the file manually.