pyston-lite: enable caching-from-type for split dicts
we need to guard on the entire keys object to make sure
that none of the attributes are the one in question. in pyston-full
we do this using the dk_version field we added, but we can't do that
in lite.

so instead guard on the value of the keys object itself plus nentries.
We have to be a little careful to make sure the keys object wasn't
deallocated + reallocated in the same location.

This results in a 2% speedup on flaskblogging and pyperformance