New LOAD_GLOBAL cache that is more permissive
It uses our OFFSET approach, where we store the offset in
the hashtable and then check to make sure the same key
is still stored there. This lets it work for cases that
the globals are being reassigned.

This new cache type is about twice as fast as no-cache, and
and the directly-cache-the-value type is twice again as fast.

Faster CPython does this by guarding on the keys version, which
we could also do (would be faster than this new cache type, and
would also allow caching builtin lookups when globals change),
but I don't think there's an easy way to get this to work for pyston-lite