lite: run bolt outside of the manylinux container
The bolt that has been built on the host does not typically work
inside the manylinux container. We could try to build bolt inside
the container, or maybe we could create a package and install it.

But it seemed easier to just generate a non-bolted wheel, then
run bolt on the host and repackage the wheel.