Make sure these are absolute paths
Bolt will silently fail to do anything meaningful if you provide a relative
path to the profiling output, since it will write all the profile data to
temporary directories that get deleted.

I ran into this when trying to port bolt to cpython, so I wanted to check
that we're using abs paths for pyston-lite. Looks like we are but I added
a call to os.path.abspath just to be sure